track the reference address of an emitter so proxy objects can work seamlessly.
If an emitter wants to masquerade as another when sending an event (by
explicitly setting the emitter attribute to the other emitter), and the
listener wants to perform introspection on itself, e.g. unsubscribing
from that emitter/event, the masqueraded emitter will not show up in
the listener's subscriptions, and the unsubscribe action will do nothing.

Now the reference address of the emitter is stored separately in an "addr"
attribute, and may be used instead of the reference when doing introspective
work.  When masquerading an event, a proxy emitter must be sure to set
the addr attribute in the event to its own, so that the listener can
identify the actual emitter it subscribed to.